mp3info/requirements.txt
chardet.py
